@@ -1,14 +1,16 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.1.8'
+gem 'rails', '4.2.0'
 
+gem 'by_star', github: 'radar/by_star'
 gem 'faker', '~> 1.4.2'
-gem 'html-pipeline', '~> 1.9.0'
+gem 'html-pipeline', '~> 1.11.0'
 gem 'kaminari', '~> 0.16.0'
 gem 'pg', '~> 0.17.0'
+gem 'pghero', '~> 0.1.0'
 gem 'pundit', '~> 0.3.0'
 gem 'redcarpet', '~> 3.2.0'
-gem 'redis', '~> 3.1.0'
+gem 'redis', '~> 3.2.0'
 gem 'redis-namespace', '~> 1.5.0'
 gem 'ruby-progressbar'
 gem 'sanitize', '~> 3.0.0'
@@ -19,25 +21,27 @@ gem 'simple_form', '~> 3.1.0'
 # Authentication
 gem 'omniauth', '~> 1.2.0'
 gem 'omniauth-openid', '1.0.1'
+gem 'omniauth-google-oauth2', '~> 0.2.6'
+gem 'omniauth-twitter', '~> 1.1.0'
 
 # Jobs & web interface
 gem 'sidekiq', '~> 3.3.0'
 gem 'sinatra', require: false
 
 # Asset gems
-gem 'bootstrap-sass', '~> 3.2.0'
+gem 'bootstrap-sass', '~> 3.3.0'
 gem 'font-awesome-sass-rails'
 gem 'jquery-rails', '~> 3.1.2'
-gem 'sass-rails',   '~> 4.0.0'
+gem 'sass-rails',   '~> 5.0.0'
 gem 'coffee-rails', '~> 4.1.0'
 gem 'uglifier', '>= 1.0.3'
 
 group :development do
   gem 'quiet_assets'
   # Guard for autorunning tests
-  gem 'guard', '~> 2.6.0'
-  gem 'guard-rspec', '~> 4.3.0'
-  gem 'guard-spinach', github: 'thomas-mcdonald/guard-spinach'
+  gem 'guard', '~> 2.10.0'
+  gem 'guard-rspec', '~> 4.4.0'
+  gem 'guard-spinach', github: 'codegram/guard-spinach'
   gem 'rails-erd', github: 'paulwittmann/rails-erd', branch: 'mavericks'
   gem 'rb-fsevent', '~> 0.9.1'
   gem 'spring'
@@ -45,6 +49,7 @@ group :development do
 end
 
 group :development, :test do
+  gem 'dotenv-rails'
   gem 'rspec-rails', '~> 3.0'
   gem 'spinach-rails'
 end
