@@ -1,5 +1,5 @@
 class AuthorizationsController < ApplicationController
-  skip_before_filter :verify_authenticity_token, only: [:callback]
+  skip_before_action :verify_authenticity_token, only: [:callback]
 
   # Acts as a callback
   def callback
@@ -8,6 +8,10 @@ class AuthorizationsController < ApplicationController
       login(@user)
       redirect_to "/" and return
     else
+      # We store the authorization so that the user cannot modify
+      # the provider/uid details in the form
+      @authorization = Authorization.create_from_hash(auth_hash)
+      session[:auth_id] = @authorization.id
       @user = User.new_from_hash(auth_hash)
     end
   end
