@@ -6,8 +6,13 @@ class Spinach::Features::Login < Spinach::FeatureSteps
     create_user
   end
 
+  step 'I open the login modal' do
+    visit '/'
+    click_link 'login-link'
+  end
+
   step 'I click on the Google provider' do
-    click_link_or_button('google-login')
+    click_link_or_button 'google-login-button'
   end
 
   step 'I should be returned to the homepage' do
