@@ -14,10 +14,7 @@ Dir["#{Rails.root}/features/steps/shared/*.rb"].each {|file| require file}
 
 # Setup Javascript tags
 require 'capybara/poltergeist'
-Capybara.javascript_driver = :poltergeist
-Spinach.hooks.on_tag("javascript") do
-  ::Capybara.current_driver = ::Capybara.javascript_driver
-end
+Capybara.default_driver = :poltergeist
 Capybara.default_wait_time = 10
 
 # Setup Omniauth mock
